+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.0.0-beta.6
 ----
  * css-components: Add San Francisco font for iOS9 devices.
  * ons-input: Add "ons-input" component and remove "ons-material-input" component.
+ * ons-range: Add "ons-range" component.
 
 v2.0.0-beta.5
 ----